feature: use custom hook useHistory to not surcharge app
