+++ interfaces/app.js
@@ -1,23 +1,17 @@
 const React = require('react')
 const { default: TerminalUI } = require('react-terminal-ui')
 
-const DefaultOutputs = require('./components/outputs')
 const Input = require('./components/input')
-const Output = require('./components/output')
-
-const { defaultTheme, switchTheme } = require('./utils/themes')
-const { APP_NAME, KEY_MODE, KEY_CLEAR, TERMINAL_HEIGHT, KEY_THEME } = require('./constants')
-
-const { fetchCmd } = require('./utils/fetch')
-
-const { getSessionStorage, setSessionStorage } = require('./utils/storage')
-
 const { _id } = require('./utils/uuid')
 
+const { defaultTheme } = require('./utils/themes')
+const { APP_NAME, TERMINAL_HEIGHT } = require('./constants')
+
+const useHistory = require('./hooks/useHistory')
+
 const App = () => {
-  const [theme, setTheme] = React.useState(getSessionStorage(KEY_THEME))
   const [inputs, setInputs] = React.useState([])
-  const [history, setHistory] = React.useState([])
+  const { history, groupHistory } = useHistory()
 
   const grpInput = (input) => {
     const ipts = [...inputs]
@@ -25,47 +19,15 @@ const App = () => {
     setInputs(ipts)
   }
 
-  const grpHistory = async (input) => {
-    let lns = [...history]
-
-    lns.push(<Input key={_id()} value={input} />)
-
-    if (input.toLocaleLowerCase().trim().slice(0, 4) === KEY_MODE) {
-      const th = switchTheme(input)
-      setTheme(th)
-      setSessionStorage(KEY_THEME, th)
-    }
-    if (input.toLocaleLowerCase().trim() === KEY_CLEAR) {
-      lns = []
-    } else if (input) {
-      const cmdRes = await fetchCmd(input)
-      lns.push(<Output key={_id()} value={cmdRes} />)
-    }
-
-    setHistory(lns)
-  }
-
-  const pushDefaultOutputs = () => {
-    const hs = [...history]
-    if (Array.isArray(history) && history.length === 0) {
-      hs.push(<DefaultOutputs key={_id()} />)
-      setHistory(hs)
-    }
-  }
-
   const handleInput = (input) => {
     grpInput(input)
-    grpHistory(input)
+    groupHistory(input)
   }
-
-  React.useEffect(() => {
-    pushDefaultOutputs()
-  }, [])
 
   return (
     <TerminalUI
       name={APP_NAME}
-      colorMode={theme ?? defaultTheme}
+      colorMode={defaultTheme}
       onInput={handleInput}
       height={TERMINAL_HEIGHT}
     >
